requirements.txt
django-allauth>=65.3.0
django>=5.1.4
django-environ>=0.11.2
django-allauth-2fa>=0.11.1
django-crispy-forms>=2.3
crispy-tailwind>=1.0.3
django-extensions>=3.2.3
gunicorn>=23.0.0
gevent>=24.11.1
